Run IterateBuiltInWebTypeTest in interactive ui tests
