+++ browser/ui/sidebar/BUILD.gn
@@ -78,6 +78,40 @@ source_set("sidebar_impl") {
   }
 }
 
+# Shared fixture (SidebarBrowserTest) consumed by both the browser_tests and
+# interactive_ui_tests source sets below. It carries the fixture's dependencies
+# as public_deps so both consumers inherit them.
+source_set("test_support") {
+  testonly = true
+
+  sources = [
+    "sidebar_browsertest_base.cc",
+    "sidebar_browsertest_base.h",
+  ]
+
+  public_deps = [
+    ":sidebar",
+    ":sidebar_impl",
+    "//base",
+    "//brave/browser/ui/views/frame",
+    "//brave/browser/ui/views/sidebar",
+    "//brave/components/sidebar/browser",
+    "//chrome/browser/ui",
+    "//chrome/browser/ui/browser_window",
+    "//chrome/browser/ui/side_panel:side_panel_views_dependent",
+    "//chrome/browser/ui/tabs:tab_strip",
+    "//chrome/test:test_support",
+    "//chrome/test:test_support_ui",
+    "//testing/gtest",
+    "//ui/events",
+    "//ui/events:test_support",
+    "//ui/gfx",
+    "//ui/gfx:test_support",
+    "//ui/views",
+    "//url",
+  ]
+}
+
 source_set("browser_tests") {
   testonly = true
   defines = [ "HAS_OUT_OF_PROC_TEST_RUNNER" ]
@@ -87,6 +121,7 @@ source_set("browser_tests") {
   deps = [
     ":sidebar",
     ":sidebar_impl",
+    ":test_support",
     "//base",
     "//brave/app:command_ids",
     "//brave/browser/ui/tabs:tabs_public",
@@ -132,6 +167,34 @@ source_set("browser_tests") {
   if (enable_playlist) {
     deps += [ "//brave/components/playlist/core/common" ]
   }
+
+  if (enable_brave_wallet) {
+    deps += [ "//brave/components/brave_wallet/common:common_constants" ]
+  }
+}
+
+source_set("interactive_ui_tests") {
+  testonly = true
+  defines = [ "HAS_OUT_OF_PROC_TEST_RUNNER" ]
+
+  sources = [ "sidebar_interactive_ui_tests.cc" ]
+
+  deps = [
+    ":sidebar_impl",
+    ":test_support",
+    "//base",
+    "//brave/components/brave_wallet/common/buildflags",
+    "//brave/components/sidebar/browser",
+    "//chrome/browser/ui",
+    "//chrome/browser/ui/browser_window",
+    "//chrome/browser/ui/tabs:tab_strip",
+    "//chrome/test:test_support",
+    "//chrome/test:test_support_ui",
+    "//content/public/browser",
+    "//content/test:test_support",
+    "//ui/base",
+    "//url",
+  ]
 
   if (enable_brave_wallet) {
     deps += [ "//brave/components/brave_wallet/common:common_constants" ]
