+++ browser/ui/sidebar/sidebar_browsertest.cc
@@ -21,6 +21,7 @@
 #include "brave/browser/ui/browser_commands.h"
 #include "brave/browser/ui/focus_mode/focus_mode_controller.h"
 #include "brave/browser/ui/focus_mode/focus_mode_features.h"
+#include "brave/browser/ui/sidebar/sidebar_browsertest_base.h"
 #include "brave/browser/ui/sidebar/sidebar_controller.h"
 #include "brave/browser/ui/sidebar/sidebar_model.h"
 #include "brave/browser/ui/sidebar/sidebar_service_factory.h"
@@ -111,215 +112,7 @@ using ::testing::Eq;
 using ::testing::Ne;
 using ::testing::Optional;
 
-namespace {
-ui::MouseEvent GetDummyEvent() {
-  return ui::MouseEvent(ui::EventType::kMouseMoved, gfx::PointF(),
-                        gfx::PointF(), base::TimeTicks::Now(), 0, 0);
-}
-}  // namespace
-
 namespace sidebar {
-
-class SidebarBrowserTest : public InProcessBrowserTest {
- public:
-  SidebarBrowserTest() = default;
-  ~SidebarBrowserTest() override = default;
-
-  void SetUpOnMainThread() override {
-    InProcessBrowserTest::SetUpOnMainThread();
-    animation_resetter_ = gfx::AnimationTestApi::SetRichAnimationRenderMode(
-        gfx::Animation::RichAnimationRenderMode::FORCE_DISABLED);
-  }
-
-  void PreRunTestOnMainThread() override {
-    InProcessBrowserTest::PreRunTestOnMainThread();
-
-    auto* service = SidebarServiceFactory::GetForProfile(browser()->profile());
-    // Enable sidebar explicitely because sidebar option is different based on
-    // channel.
-    service->SetSidebarShowOption(
-        SidebarService::ShowSidebarOption::kShowAlways);
-  }
-
-  SidebarModel* model() const { return controller()->model(); }
-  TabStripModel* tab_model() const { return browser()->tab_strip_model(); }
-
-  SidebarController* controller() const {
-    return browser()->GetFeatures().sidebar_controller();
-  }
-
-  SidePanelButton* GetSidePanelToolbarButton() const {
-    return static_cast<BraveToolbarView*>(
-               BrowserView::GetBrowserViewForBrowser(browser())->toolbar())
-        ->side_panel_button();
-  }
-
-  views::View* GetVerticalTabsContainer() const {
-    auto* view = BrowserView::GetBrowserViewForBrowser(browser());
-    return static_cast<BraveBrowserView*>(view)->vertical_tab_strip_host_view_;
-  }
-
-  raw_ptr<SidebarItemsContentsView> GetSidebarItemsContentsView(
-      SidebarController* controller) const {
-    auto* sidebar_container_view =
-        static_cast<SidebarContainerView*>(controller->sidebar());
-    auto sidebar_control_view = sidebar_container_view->sidebar_control_view_;
-    auto sidebar_scroll_view = sidebar_control_view->sidebar_items_view_;
-    auto sidebar_items_contents_view = sidebar_scroll_view->contents_view_;
-    DCHECK(sidebar_items_contents_view);
-
-    return sidebar_items_contents_view;
-  }
-
-  SidebarItemsScrollView* GetSidebarItemsScrollView(
-      SidebarController* controller) const {
-    auto* sidebar_container_view =
-        static_cast<SidebarContainerView*>(controller->sidebar());
-    auto sidebar_control_view = sidebar_container_view->sidebar_control_view_;
-    return sidebar_control_view->sidebar_items_view_;
-  }
-
-  // If the item at |index| is panel item, this will return after waiting
-  // model's active index is changed as active index could not be not updated
-  // synchronously. Panel activation is done via SidePanelCoordinator instead of
-  // asking activation to SidebarController directly.
-  void SimulateSidebarItemClickAt(size_t index) {
-    auto sidebar_items_contents_view =
-        GetSidebarItemsContentsView(controller());
-
-    auto* item = sidebar_items_contents_view->children()[index].get();
-    DCHECK(item);
-
-    const gfx::Point origin(0, 0);
-    ui::MouseEvent event(ui::EventType::kMousePressed, origin, origin,
-                         ui::EventTimeForNow(), ui::EF_LEFT_MOUSE_BUTTON, 0);
-    sidebar_items_contents_view->OnItemPressed(item, event);
-
-    if (model()->GetAllSidebarItems()[index].open_in_panel) {
-      auto* panel_ui = browser()->GetFeatures().side_panel_ui();
-      WaitUntil(base::BindLambdaForTesting([&]() {
-        return (model()->active_index() == index &&
-                panel_ui->IsSidePanelShowing());
-      }));
-    }
-  }
-
-  void HandleBrowserWindowMouseEvent() {
-    browser_view()->HandleBrowserWindowMouseEvent(GetDummyEvent());
-  }
-
-  SidebarControlView* GetSidebarControlView() const {
-    return GetSidebarContainerView()->sidebar_control_view_;
-  }
-
-  SidebarContainerView* GetSidebarContainerView() const {
-    return static_cast<SidebarContainerView*>(controller()->sidebar());
-  }
-
-  SidePanel* GetSidePanel() { return browser_view()->side_panel(); }
-
-  bool IsSidebarUIOnLeft() const {
-    return GetSidebarContainerView()->sidebar_on_left_ &&
-           GetSidebarControlView()->sidebar_on_left_;
-  }
-
-  void ShowSidebar() { GetSidebarContainerView()->ShowSidebar(); }
-
-  void HideSidebar() { GetSidebarContainerView()->HideSidebar(); }
-
-  void WaitUntil(base::RepeatingCallback<bool()> condition) {
-    if (condition.Run()) {
-      return;
-    }
-
-    base::RepeatingTimer scheduler;
-    scheduler.Start(FROM_HERE, base::Milliseconds(100),
-                    base::BindLambdaForTesting([this, &condition] {
-                      if (condition.Run()) {
-                        run_loop_->Quit();
-                      }
-                    }));
-    Run();
-  }
-
-  void Run() {
-    run_loop_ = std::make_unique<base::RunLoop>();
-    run_loop()->Run();
-  }
-
-  void SetItemAddedBubbleLaunchedCallback(
-      SidebarItemsContentsView* items_contents_view) {
-    items_contents_view->item_added_bubble_launched_for_test_ =
-        base::BindRepeating(
-            &SidebarBrowserTest::ItemAddedBubbleLaunchedCallback,
-            weak_factory_.GetWeakPtr());
-  }
-
-  void ItemAddedBubbleLaunchedCallback(views::View* anchor) {
-    item_added_bubble_anchor_ = anchor;
-  }
-
-  void AddItemsTillScrollable(SidebarItemsScrollView* scroll_view,
-                              SidebarService* sidebar_service) {
-    int url_prefix = 0;
-    while (true) {
-      sidebar_service->AddItem(sidebar::SidebarItem::Create(
-          GURL(base::StrCat(
-              {"https://foo/bar_", base::NumberToString(url_prefix)})),
-          u"title", SidebarItem::Type::kTypeWeb,
-          SidebarItem::BuiltInItemType::kNone, false));
-      url_prefix++;
-      base::RunLoop().RunUntilIdle();
-      // Add items till first item becomes invisible.
-      if (scroll_view->NeedScrollForItemAt(0)) {
-        break;
-      }
-    }
-  }
-
-  bool NeedScrollForItemAt(size_t index, SidebarItemsScrollView* scroll_view) {
-    return scroll_view->NeedScrollForItemAt(index);
-  }
-
-  void VerifyTargetDragIndicatorIndexCalc(const gfx::Point& screen_position) {
-    auto sidebar_items_contents_view =
-        GetSidebarItemsContentsView(controller());
-    EXPECT_NE(std::nullopt,
-              sidebar_items_contents_view->CalculateTargetDragIndicatorIndex(
-                  screen_position));
-  }
-
-  base::RunLoop* run_loop() const { return run_loop_.get(); }
-
-  size_t GetDefaultItemCount() const {
-    return SidebarServiceFactory::GetForProfile(browser()->profile())
-        ->GetDefaultSidebarItemsCountForTesting();
-  }
-
-  int GetFirstPanelItemIndex() {
-    auto const items = model()->GetAllSidebarItems();
-    auto const iter =
-        std::ranges::find(items, true, &SidebarItem::open_in_panel);
-    return std::distance(items.cbegin(), iter);
-  }
-
-  int GetFirstWebItemIndex() {
-    const auto items = model()->GetAllSidebarItems();
-    auto const iter =
-        std::ranges::find(items, false, &SidebarItem::open_in_panel);
-    return std::distance(items.cbegin(), iter);
-  }
-
-  BraveBrowserView* browser_view() {
-    return BraveBrowserView::From(
-        BrowserView::GetBrowserViewForBrowser(browser()));
-  }
-
-  raw_ptr<views::View, DanglingUntriaged> item_added_bubble_anchor_ = nullptr;
-  std::unique_ptr<base::RunLoop> run_loop_;
-  gfx::AnimationTestApi::RenderModeResetter animation_resetter_;
-  base::WeakPtrFactory<SidebarBrowserTest> weak_factory_{this};
-};
 
 IN_PROC_BROWSER_TEST_F(SidebarBrowserTest, BasicTest) {
   EXPECT_TRUE(!!GetSidePanelToolbarButton()->context_menu_controller());
@@ -470,77 +263,6 @@ IN_PROC_BROWSER_TEST_F(SidebarBrowserTest, WebTypePanelTest) {
   // New tab is not created.
   EXPECT_EQ(2, tab_model()->count());
 }
-
-#if BUILDFLAG(ENABLE_BRAVE_WALLET)
-IN_PROC_BROWSER_TEST_F(SidebarBrowserTest, IterateBuiltInWebTypeTest) {
-  // Click builtin wallet item and it's loaded at current active tab.
-  const auto items = model()->GetAllSidebarItems();
-  const auto wallet_item_iter =
-      std::ranges::find(items, SidebarItem::BuiltInItemType::kWallet,
-                        &SidebarItem::built_in_item_type);
-  ASSERT_NE(wallet_item_iter, items.cend());
-  const int wallet_item_index = std::distance(items.cbegin(), wallet_item_iter);
-  auto wallet_item = model()->GetAllSidebarItems()[wallet_item_index];
-  EXPECT_FALSE(controller()->DoesBrowserHaveOpenedTabForItem(wallet_item));
-  SimulateSidebarItemClickAt(wallet_item_index);
-  EXPECT_TRUE(controller()->DoesBrowserHaveOpenedTabForItem(wallet_item));
-  EXPECT_EQ(0, tab_model()->active_index());
-  EXPECT_EQ(tab_model()->GetWebContentsAt(0)->GetVisibleURL().host(),
-            wallet_item.url.host());
-
-  // Create NTP and click wallet item. Then wallet tab(index 0) is activated.
-  ASSERT_TRUE(ui_test_utils::NavigateToURLWithDisposition(
-      browser(), GURL("brave://newtab/"),
-      WindowOpenDisposition::NEW_FOREGROUND_TAB,
-      ui_test_utils::BROWSER_TEST_WAIT_FOR_LOAD_STOP));
-  // NTP is active tab.
-  EXPECT_EQ(1, tab_model()->active_index());
-  SimulateSidebarItemClickAt(wallet_item_index);
-  // Wallet tab is active tab.
-  EXPECT_EQ(0, tab_model()->active_index());
-  EXPECT_EQ(tab_model()->GetWebContentsAt(0)->GetVisibleURL().host(),
-            wallet_item.url.host());
-
-  // Create NTP.
-  ASSERT_TRUE(ui_test_utils::NavigateToURLWithDisposition(
-      browser(), GURL("brave://newtab/"),
-      WindowOpenDisposition::NEW_FOREGROUND_TAB,
-      ui_test_utils::BROWSER_TEST_WAIT_FOR_LOAD_STOP));
-  // NTP is active tab and load wallet on it.
-  EXPECT_EQ(2, tab_model()->active_index());
-  ASSERT_TRUE(ui_test_utils::NavigateToURL(browser(), wallet_item.url));
-
-  // Click wallet item and then first wallet tab(tab index 0) is activated.
-  SimulateSidebarItemClickAt(wallet_item_index);
-  EXPECT_EQ(0, tab_model()->active_index());
-
-  // Click wallet item and then second wallet tab(index 2) is activated.
-  SimulateSidebarItemClickAt(wallet_item_index);
-  EXPECT_EQ(2, tab_model()->active_index());
-
-  // Click wallet item and then first wallet tab(index 0) is activated.
-  SimulateSidebarItemClickAt(wallet_item_index);
-  EXPECT_EQ(0, tab_model()->active_index());
-
-  // Checking windows' activation state is flaky in browser tests.
-#if !BUILDFLAG(IS_MAC)
-  auto* browser2 = CreateBrowser(browser()->profile());
-  WaitUntil(base::BindLambdaForTesting(
-      [&]() { return browser2->window()->IsActive(); }));
-
-  // |browser2| doesn't have any wallet tab. So, clicking wallet sidebar item
-  // activates other browser's first wallet tab.
-  browser2->GetFeatures().sidebar_controller()->ActivateItemAt(
-      wallet_item_index);
-
-  // Wait till browser() is activated.
-  WaitUntil(base::BindLambdaForTesting(
-      [&]() { return browser()->window()->IsActive(); }));
-
-  EXPECT_EQ(0, tab_model()->active_index());
-#endif
-}
-#endif  // BUILDFLAG(ENABLE_BRAVE_WALLET)
 
 IN_PROC_BROWSER_TEST_F(SidebarBrowserTest, PRE_LastlyUsedSidePanelItemTest) {
   auto* panel_ui = browser()->GetFeatures().side_panel_ui();
