+++ test/BUILD.gn
@@ -810,6 +810,7 @@ if (!is_android) {
     deps = [
       ":browser_test_support",
       "//brave:packed_resources",
+      "//brave/browser/ui/sidebar:interactive_ui_tests",
       "//chrome:packed_resources",
       "//chrome/browser",
       "//chrome/test:chrome_test_launcher",
